changing details in the methods : the division is now a double, not an int, and the coordinates to teleport are not between 1000 and -1000

+++ src/main/java/eu/pommeray/randomadditions/commands/Calculator.java
@@ -66,7 +66,7 @@ public class Calculator implements CommandExecutor {
       }
       case "/": {
         if (secondNumber != 0) {
-          int result = firstNumber / secondNumber;
+          double result = (double) firstNumber / secondNumber;
           sender.sendMessage("The result of " + firstNumber
                       + " / " + secondNumber + " is: " + result);
         } else {
